US09 - Gestão Administrativa de Organizações e Usuários (#40)
* criando endpoint de grupo por id e ajustando listagem de grupos

* criando endpoint de delete grupo por id

* implementando get, put e delete de pesquisadores/usuario

* ajustando permissoes de role

+++ src/main/java/com/orbytum/api/fachada/AuthFachada.java
@@ -93,16 +93,22 @@ public class AuthFachada {
         );
         novoAdmin = usuarioService.save(novoAdmin);
 
+        Usuario criador = (credenciaisSolicitante.getAccessLevel() == AccessLevel.INITIAL_ADMIN)
+                ? null
+                : credenciaisSolicitante.getUsuario();
+
         CredenciaisLogin credenciaisAdmin = new CredenciaisLogin(
                 request.email(),
                 passwordEncoder.encode(request.senha()),
                 AccessLevel.ADMIN,
                 novoAdmin,
-                credenciaisSolicitante.getUsuario()
+                criador
         );
         credenciaisLoginService.save(credenciaisAdmin);
 
-        excluirAdminInicialSeExistir();
+        if (credenciaisSolicitante.getAccessLevel() == AccessLevel.INITIAL_ADMIN) {
+            excluirAdminInicialSeExistir();
+        }
 
         UserDetails userDetails = User.builder()
                 .username(request.email())
